requirements.txt
Django==3.1.2
djangorestframework==3.12.1
requests==2.24.0
clearbit
